+++ core/src/main/scala/com/klibisz/elastiknn/models/HashingFunction.scala
@@ -1,0 +1,9 @@
+package com.klibisz.elastiknn.models
+
+import com.klibisz.elastiknn.api.{Mapping, Vec}
+import com.klibisz.elastiknn.storage.StoredVec
+
+trait HashingFunction[M <: Mapping, V <: Vec, S <: StoredVec] extends (V => Array[Array[Byte]]) {
+  val mapping: M
+  val exact: ExactSimilarityFunction[V, S]
+}